requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
redis
google-generativeai
beautifulsoup4
requests
chromadb
